aider: Add functionality to check for the existence of config.yaml and offer to copy config.yaml.example if it doesn't exist.
# Aider chat conversation:

USER: Have the script look for config.yaml - if it doesn't exist, have the script offer to copy `config.yaml.example` to config.yaml

ASSISTANT: To implement this, we need to modify `movephoto.go` to:

1. Check if `config.yaml` exists.
2. If it doesn't exist, offer to copy `config.yaml.example` to `config.yaml`.

Here is the *SEARCH/REPLACE* block:

movephoto.go
```go
<<<<<<< SEARCH
func main() {
	purge_unwanted(robertWatchDir, bannedExtensions)
	move_photos(robertWatchDir, defaultDestinationDir, imageExtensions)
	move_videos(robertWatchDir, defaultDestinationDir)
}
=======
import (
	"os"
	"fmt"
	"bufio"
)

func main() {
	if _, err := os.Stat("config.yaml"); os.IsNotExist(err) {
		fmt.Println("config.yaml does not exist. Would you like to copy config.yaml.example to config.yaml? (y/n)")
		reader := bufio.NewReader(os.Stdin)
		response, _ := reader.ReadString('\n')
		if response == "y\n" {
			os.Link("config.yaml.example", "config.yaml")
		}
	}
	purge_unwanted(robertWatchDir, bannedExtensions)
	move_photos(robertWatchDir, defaultDestinationDir, imageExtensions)
	move_videos(robertWatchDir, defaultDestinationDir)
}
>>>>>>> REPLACE
```

Please note that this code will only copy the file if the user types 'y' and hits enter when prompted. If they type anything else, the file will not be copied.