aider: Added information on configuration to readme.md. If config.yaml didn't exist, it offered to copy the contents of config.yaml.example as a template, and if it needed to be created, the script would exit with a message asking the user to review the contents of config.yaml before running the script again.
# Aider chat conversation:

USER: Add information on configuration to readme.md. Explain that if a config.yaml file doesn't exist, it will offer to copy the contents of config.yaml.example as a template. If config.yaml needed to be created, ensure that the script will exit with a message asking the user to review the contents of config.yaml before running the script again.

ASSISTANT: To add this information to the `readme.md` file, we need to:

1. Add a section explaining that if a `config.yaml` file doesn't exist, the script will offer to copy the contents of `config.yaml.example` as a template.
2. Add a section explaining that if `config.yaml` needed to be created, the script will exit with a message asking the user to review the contents of `config.yaml` before running the script again.

Here are the *SEARCH/REPLACE* blocks:

readme.md
<source>
<<<<<<< SEARCH
## Configuration

The script is configured using a `config.yaml` file. An example configuration file `config.yaml.example` is provided. Copy this file to `config.yaml` and modify it to suit your needs.
=======
## Configuration

The script is configured using a `config.yaml` file. If a `config.yaml` file doesn't exist when the script is run, it will offer to copy the contents of `config.yaml.example` as a template.

An example configuration file `config.yaml.example` is provided. Copy this file to `config.yaml` and modify it to suit your needs. If `config.yaml` needed to be created, the script will exit with a message asking the user to review the contents of `config.yaml` before running the script again.
>>>>>>> REPLACE
</source>