aider: The script now performs a single scan if --watch isn't added and replaces fsnotify with regular scans of the source directories when --watch is used.
# Aider chat conversation:

USER: I ran this command:

 ./movephoto --watch -debug

And got this output:

flag provided but not defined: -watch
Usage of ./movephoto:
  -debug
    	Enable debug output
  -polling-interval int
    	Polling interval in seconds for checking new files in the watch directories (default 30)

ASSISTANT: The error message indicates that the `--watch` flag is not defined in the `movephoto.go` program. Since you have previously decided to remove filesystem watching and go back to a single-threaded polling system, the `--watch` flag should no longer be used. Instead, the program now uses a polling system with a configurable interval, and the `--debug` flag is available for enabling debug output.

To resolve this issue, you should run the program without the `--watch` flag, as it is no longer supported. If you want to enable debug output, make sure to use the correct flag syntax with two dashes (`--debug`). Here's the correct command to run the program with debug output enabled:

```sh
./movephoto --debug
```

If you need to adjust the polling interval, you can use the `--polling-interval` flag followed by the number of seconds you want to set as the interval. For example, to set the polling interval to 60 seconds with debug output, you would run:

```sh
./movephoto --debug --polling-interval 60
```

If you need further assistance or if there are additional features or changes you would like to implement, please let me know.

USER: I want the script to continue to scan that directory. However I want to replace fsnotify with regular scans of the source directories. I want the script to do a single scan if --watch isn't added

ASSISTANT: To modify the `movephoto.go` script to perform a single scan when the `--watch` flag isn't provided and to replace `fsnotify` with regular scans of the source directories when the `--watch` flag is used, we need to make the following changes:

1. Add a `watch` boolean flag to enable or disable the watching functionality.
2. Modify the `main` function to check the `watch` flag and perform a single scan if it's not set, or start the regular scanning process if it is set.

Here are the *SEARCH/REPLACE* blocks for these changes:

movephoto.go
```go
<<<<<<< SEARCH
var (
	pollingInterval = flag.Int("polling-interval", 30, "Polling interval in seconds for checking new files in the watch directories")
	debug         = flag.Bool("debug", false, "Enable debug output")
)
=======
var (
	pollingInterval = flag.Int("polling-interval", 30, "Polling interval in seconds for checking new files in the watch directories")
	debug           = flag.Bool("debug", false, "Enable debug output")
	watch           = flag.Bool("watch", false, "Enable regular scanning of the source directories")
)
>>>>>>> REPLACE
```

movephoto.go
```go
<<<<<<< SEARCH
func main() {
	flag.Parse() // Parse the command-line flags

	if *debug {
		log.Printf("[%s] Starting movephoto...\n", currentTime())
	}
	config := loadConfig()

	processFiles(config)

	ticker := time.NewTicker(time.Duration(*pollingInterval) * time.Second)
	defer ticker.Stop()

	for range ticker.C {
		if *debug {
			log.Printf("[%s] Polling for new files...\n", currentTime())
		}
		processFiles(config)
	}
}
=======
func main() {
	flag.Parse() // Parse the command-line flags

	if *debug {
		log.Printf("[%s] Debug mode enabled\n", currentTime())
	}

	config := loadConfig()

	if !*watch {
		if *debug {
			log.Printf("[%s] Performing a single scan...\n", currentTime())
		}
		processFiles(config)
	} else {
		if *debug {
			log.Printf("[%s] Starting regular scanning of source directories...\n", currentTime())
		}
		ticker := time.NewTicker(time.Duration(*pollingInterval) * time.Second)
		defer ticker.Stop()

		for range ticker.C {
			if *debug {
				log.Printf("[%s] Polling for new files...\n", currentTime())
			}
			processFiles(config)
		}
	}
}
>>>>>>> REPLACE
```

Please apply these changes to the `movephoto.go` file. After making these changes, running the script with `--watch` will start regular scans of the source directories, and running it without any flags will perform a single scan of the directories.